+++ collections/src/main/kotlin/ktx/collections/maps.kt
@@ -5,6 +5,9 @@ package ktx.collections
 import com.badlogic.gdx.utils.*
 import com.badlogic.gdx.utils.ObjectMap.Entry
 
+/** Alias for [com.badlogic.gdx.utils.ObjectMap]. Added for consistency with other collections and factory methods. */
+typealias GdxMap<Key, Value> = ObjectMap<Key, Value>
+
 /**
  * Default LibGDX map size used by most constructors.
  */
@@ -15,8 +18,8 @@ const val defaultMapSize = 51
  * @param loadFactor decides under what load the map is resized.
  * @return a new [ObjectMap].
  */
-fun <Key, Value> gdxMapOf(initialCapacity: Int = defaultSetSize, loadFactor: Float = defaultLoadFactor): ObjectMap<Key, Value> =
-    ObjectMap(initialCapacity, loadFactor)
+fun <Key, Value> gdxMapOf(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor): GdxMap<Key, Value> =
+    GdxMap(initialCapacity, loadFactor)
 
 /**
  * @param keysToValues will be added to the map.
@@ -25,9 +28,9 @@ fun <Key, Value> gdxMapOf(initialCapacity: Int = defaultSetSize, loadFactor: Flo
  * @return a new [ObjectMap].
  */
 inline fun <Key, Value> gdxMapOf(vararg keysToValues: Pair<Key, Value>,
-                                 initialCapacity: Int = defaultSetSize,
-                                 loadFactor: Float = defaultLoadFactor): ObjectMap<Key, Value> {
-  val map = ObjectMap<Key, Value>(initialCapacity, loadFactor)
+                                 initialCapacity: Int = defaultMapSize,
+                                 loadFactor: Float = defaultLoadFactor): GdxMap<Key, Value> {
+  val map = GdxMap<Key, Value>(initialCapacity, loadFactor)
   keysToValues.forEach { map[it.first] = it.second }
   return map
 }
@@ -36,30 +39,30 @@ inline fun <Key, Value> gdxMapOf(vararg keysToValues: Pair<Key, Value>,
  * A method wrapper over [ObjectMap.size] variable compatible with nullable types.
  * @return current amount of elements in the map.
  */
-inline fun ObjectMap<*, *>?.size(): Int = this?.size ?: 0
+inline fun GdxMap<*, *>?.size(): Int = this?.size ?: 0
 
 /**
  * @return true if the map is null or has no elements.
  */
-inline fun ObjectMap<*, *>?.isEmpty(): Boolean = this == null || this.size == 0
+inline fun GdxMap<*, *>?.isEmpty(): Boolean = this == null || this.size == 0
 
 /**
  * @return true if the map is not null and contains at least one element.
  */
-inline fun ObjectMap<*, *>?.isNotEmpty(): Boolean = this != null && this.size > 0
+inline fun GdxMap<*, *>?.isNotEmpty(): Boolean = this != null && this.size > 0
 
 /**
  * @param key a value might be assigned to this key and stored in the map.
  * @return true if a value is associated with passed key. False otherwise.
  */
-operator fun <Key> ObjectMap<Key, *>.contains(key: Key): Boolean = this.containsKey(key)
+operator fun <Key> GdxMap<Key, *>.contains(key: Key): Boolean = this.containsKey(key)
 
 /**
  * @param key the passed value will be linked with this key.
  * @param value will be stored in the map, accessible by the passed key.
  * @return old value associated with the key or null if none.
  */
-operator fun <Key, Value> ObjectMap<Key, Value>.set(key: Key, value: Value): Value? = this.put(key, value)
+operator fun <Key, Value> GdxMap<Key, Value>.set(key: Key, value: Value): Value? = this.put(key, value)
 
 /**
  * Allows to iterate over the map with Kotlin lambda syntax and direct access to [MutableIterator], which can remove
@@ -67,7 +70,7 @@ operator fun <Key, Value> ObjectMap<Key, Value>.set(key: Key, value: Value): Val
  * @param action will be invoked on each key and value pair. Passed iterator is ensured to be the same instance throughout
  *    the iteration. It can be used to remove elements.
  */
-inline fun <Key, Value> ObjectMap<Key, Value>.iterate(action: (Key, Value, MutableIterator<Entry<Key, Value>>) -> Unit) {
+inline fun <Key, Value> GdxMap<Key, Value>.iterate(action: (Key, Value, MutableIterator<Entry<Key, Value>>) -> Unit) {
   val iterator = this.iterator()
   while (iterator.hasNext) {
     val next = iterator.next()
@@ -78,7 +81,7 @@ inline fun <Key, Value> ObjectMap<Key, Value>.iterate(action: (Key, Value, Mutab
 /**
  * @return keys from this map stored in an [ObjectSet].
  */
-fun <Key> ObjectMap<Key, *>.toGdxSet(): ObjectSet<Key> = this.keys().toGdxSet()
+fun <Key> GdxMap<Key, *>.toGdxSet(): ObjectSet<Key> = this.keys().toGdxSet()
 
 /**
  * @param initialCapacity initial capacity of the map. Will be resized if necessary.
@@ -87,8 +90,8 @@ fun <Key> ObjectMap<Key, *>.toGdxSet(): ObjectSet<Key> = this.keys().toGdxSet()
  * @return values copied from this iterable stored in a LibGDX map, mapped to the keys returned by the provider.
  */
 inline fun <Key, Value> Iterable<Value>.toGdxMap(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor,
-                                                 keyProvider: (Value) -> Key): ObjectMap<Key, Value> {
-  val map = ObjectMap<Key, Value>(initialCapacity, loadFactor)
+                                                 keyProvider: (Value) -> Key): GdxMap<Key, Value> {
+  val map = GdxMap<Key, Value>(initialCapacity, loadFactor)
   this.forEach { map[keyProvider(it)] = it }
   return map
 }
@@ -103,8 +106,8 @@ inline fun <Key, Value> Iterable<Value>.toGdxMap(initialCapacity: Int = defaultM
 inline fun <Type, Key, Value> Iterable<Type>.toGdxMap(initialCapacity: Int = defaultMapSize,
                                                       loadFactor: Float = defaultLoadFactor,
                                                       valueProvider: (Type) -> Value,
-                                                      keyProvider: (Type) -> Key): ObjectMap<Key, Value> {
-  val map = ObjectMap<Key, Value>(initialCapacity, loadFactor)
+                                                      keyProvider: (Type) -> Key): GdxMap<Key, Value> {
+  val map = GdxMap<Key, Value>(initialCapacity, loadFactor)
   this.forEach { map[keyProvider(it)] = valueProvider(it) }
   return map
 }
@@ -116,8 +119,8 @@ inline fun <Type, Key, Value> Iterable<Type>.toGdxMap(initialCapacity: Int = def
  * @return values copied from this array stored in a LibGDX map, mapped to the keys returned by the provider.
  */
 inline fun <Key, Value> Array<Value>.toGdxMap(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor,
-                                              keyProvider: (Value) -> Key): ObjectMap<Key, Value> {
-  val map = ObjectMap<Key, Value>(initialCapacity, loadFactor)
+                                              keyProvider: (Value) -> Key): GdxMap<Key, Value> {
+  val map = GdxMap<Key, Value>(initialCapacity, loadFactor)
   this.forEach { map[keyProvider(it)] = it }
   return map
 }
@@ -132,8 +135,8 @@ inline fun <Key, Value> Array<Value>.toGdxMap(initialCapacity: Int = defaultMapS
 inline fun <Type, Key, Value> Array<Type>.toGdxMap(initialCapacity: Int = defaultMapSize,
                                                    loadFactor: Float = defaultLoadFactor,
                                                    valueProvider: (Type) -> Value,
-                                                   keyProvider: (Type) -> Key): ObjectMap<Key, Value> {
-  val map = ObjectMap<Key, Value>(initialCapacity, loadFactor)
+                                                   keyProvider: (Type) -> Key): GdxMap<Key, Value> {
+  val map = GdxMap<Key, Value>(initialCapacity, loadFactor)
   this.forEach { map[keyProvider(it)] = valueProvider(it) }
   return map
 }
@@ -146,7 +149,7 @@ inline fun <Type, Key, Value> Array<Type>.toGdxMap(initialCapacity: Int = defaul
  * @param loadFactor decides under what load the map is resized.
  * @return a new [IdentityMap], which compares keys by references.
  */
-fun <Key, Value> gdxIdentityMapOf(initialCapacity: Int = defaultSetSize, loadFactor: Float = defaultLoadFactor):
+fun <Key, Value> gdxIdentityMapOf(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor):
     IdentityMap<Key, Value> = IdentityMap(initialCapacity, loadFactor)
 
 /**
@@ -156,7 +159,7 @@ fun <Key, Value> gdxIdentityMapOf(initialCapacity: Int = defaultSetSize, loadFac
  * @return a new [IdentityMap], which compares keys by references.
  */
 inline fun <Key, Value> gdxIdentityMapOf(vararg keysToValues: Pair<Key, Value>,
-                                         initialCapacity: Int = defaultSetSize,
+                                         initialCapacity: Int = defaultMapSize,
                                          loadFactor: Float = defaultLoadFactor): IdentityMap<Key, Value> {
   val map = IdentityMap<Key, Value>(initialCapacity, loadFactor)
   keysToValues.forEach { map[it.first] = it.second }
@@ -197,7 +200,7 @@ inline fun <Key, Value> IdentityMap<Key, Value>.iterate(action: (Key, Value, Mut
  * @param loadFactor decides under what load the map is resized.
  * @return a new [IntIntMap] with primitive int keys and values.
  */
-fun gdxIntIntMap(initialCapacity: Int = defaultSetSize, loadFactor: Float = defaultLoadFactor): IntIntMap
+fun gdxIntIntMap(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor): IntIntMap
     = IntIntMap(initialCapacity, loadFactor)
 
 /**
@@ -225,7 +228,7 @@ operator fun IntIntMap.get(key: Int): Int = this.get(key, 0)
  * @param loadFactor decides under what load the map is resized.
  * @return a new [IntFloatMap] with primitive int keys and primitive float values.
  */
-fun gdxIntFloatMap(initialCapacity: Int = defaultSetSize, loadFactor: Float = defaultLoadFactor): IntFloatMap
+fun gdxIntFloatMap(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor): IntFloatMap
     = IntFloatMap(initialCapacity, loadFactor)
 
 /**
@@ -253,7 +256,7 @@ operator fun IntFloatMap.get(key: Int): Float = this.get(key, 0f)
  * @param loadFactor decides under what load the map is resized.
  * @return a new [IntMap] with primitive int keys.
  */
-fun <Value> gdxIntMap(initialCapacity: Int = defaultSetSize, loadFactor: Float = defaultLoadFactor): IntMap<Value>
+fun <Value> gdxIntMap(initialCapacity: Int = defaultMapSize, loadFactor: Float = defaultLoadFactor): IntMap<Value>
     = IntMap(initialCapacity, loadFactor)
 
 /**